refactor: rename user type imports and update role handling in user components
