+++ src/app/(admin)/users/_components/UsersColumn.tsx
@@ -3,7 +3,7 @@
 import { ColumnDef } from "@tanstack/react-table";
 
 import { DataTableColumnHeader } from "@/shared/components/data-table/DataTableColumnHeaderProps";
-import { User } from "../_types/user";
+import { User } from "../_types/user.types";
 import { UsersTableActions } from "./UsersTablesActions";
 
 // Esta función crea las columnas y recibe una función para manejar la expansión
@@ -38,7 +38,7 @@ export const columnsUsers = (): ColumnDef<User>[] => [
     accessorKey: "roles",
     header: ({ column }) => <DataTableColumnHeader column={column} title="Rol" />,
     cell: ({ row }) => {
-      return <div className=" capitalize">{row.original.roleIds.map((roleId) => roleId).join(", ")}</div>;
+      return <div className=" capitalize">{row.original.roles.map((role) => role.name).join(", ")}</div>;
     },
   },
   {
