+++ src/app/(admin)/users/_types/user.types.ts
@@ -1,12 +1,8 @@
 import { components } from "@/lib/api/types/api";
 
 // Tipo original de la API
-type OriginalUser = components["schemas"]["UserResponseDto"];
-
-// Tipo modificado con permissionIds como string[] en lugar de string[][]
-export type User = Omit<OriginalUser, "roleIds"> & {
-  roleIds: string[];
-};
+export type UserResponseOriginal = components["schemas"]["UserResponseDto"];
+export type User = UserResponseOriginal & { roles: Role[] };
 
 export type UserCreate = Omit<components["schemas"]["CreateUserDto"], "roleIds"> & {
   roleIds: string[];
